create identity package
Factor out code from policy package related to identity allocation, types, etc.
into a separate package. This was motivated by cyclic import issues faced in
PR #2875. Update code to use this package accordingly. No change in
functionality should occur as part of this commit.

Signed-off by: Ian Vernon <ian@cilium.io>
