+++ daemon/identity.go
@@ -17,9 +17,9 @@ package main
 import (
 	"github.com/cilium/cilium/api/v1/models"
 	. "github.com/cilium/cilium/api/v1/server/restapi/policy"
+	"github.com/cilium/cilium/pkg/identity"
 	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/logging/logfields"
-	"github.com/cilium/cilium/pkg/policy"
 
 	"github.com/go-openapi/runtime/middleware"
 )
@@ -35,9 +35,9 @@ func (h *getIdentity) Handle(params GetIdentityParams) middleware.Responder {
 	if params.Labels == nil {
 		// if labels is nil, return all identities from the kvstore
 		// This is in response to "identity list" command
-		identities = policy.GetIdentities()
+		identities = identity.GetIdentities()
 	} else {
-		identity := policy.LookupIdentity(labels.NewLabelsFromModel(params.Labels))
+		identity := identity.LookupIdentity(labels.NewLabelsFromModel(params.Labels))
 		if identity == nil {
 			return NewGetIdentityIDNotFound()
 		}
@@ -53,12 +53,12 @@ type getIdentityID struct{}
 func newGetIdentityIDHandler(d *Daemon) GetIdentityIDHandler { return &getIdentityID{} }
 
 func (h *getIdentityID) Handle(params GetIdentityIDParams) middleware.Responder {
-	nid, err := policy.ParseNumericIdentity(params.ID)
+	nid, err := identity.ParseNumericIdentity(params.ID)
 	if err != nil {
 		return NewGetIdentityIDBadRequest()
 	}
 
-	identity := policy.LookupIdentityByID(nid)
+	identity := identity.LookupIdentityByID(nid)
 	if identity == nil {
 		return NewGetIdentityIDNotFound()
 	}
