+++ pkg/identity/numericidentity.go
@@ -1,0 +1,109 @@
+// Copyright 2016-2018 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package identity
+
+import (
+	"strconv"
+	"time"
+
+	"github.com/cilium/cilium/pkg/labels"
+)
+
+const (
+	secLabelTimeout = 120 * time.Second
+
+	// MinimalNumericIdentity represents the minimal numeric identity not
+	// used for reserved purposes.
+	MinimalNumericIdentity = NumericIdentity(256)
+
+	// InvalidIdentity is the identity assigned if the identity is invalid
+	// or not determined yet
+	InvalidIdentity = NumericIdentity(0)
+)
+
+const (
+	// IdentityUnknown represents an unknown identity
+	IdentityUnknown NumericIdentity = iota
+
+	// ReservedIdentityHost represents the local host
+	ReservedIdentityHost
+
+	// ReservedIdentityWorld represents any endpoint outside of the cluster
+	ReservedIdentityWorld
+
+	// ReservedIdentityCluster represents any endpoint inside the cluster
+	// that does not have a more specific identity
+	ReservedIdentityCluster
+
+	// ReservedIdentityHealth represents the local cilium-health endpoint
+	ReservedIdentityHealth
+)
+
+var (
+	ReservedIdentities = map[string]NumericIdentity{
+		labels.IDNameHost:    ReservedIdentityHost,
+		labels.IDNameWorld:   ReservedIdentityWorld,
+		labels.IDNameHealth:  ReservedIdentityHealth,
+		labels.IDNameCluster: ReservedIdentityCluster,
+	}
+	ReservedIdentityNames = map[NumericIdentity]string{
+		ReservedIdentityHost:    labels.IDNameHost,
+		ReservedIdentityWorld:   labels.IDNameWorld,
+		ReservedIdentityHealth:  labels.IDNameHealth,
+		ReservedIdentityCluster: labels.IDNameCluster,
+	}
+)
+
+// NumericIdentity is the numeric representation of a security identity / a
+// security policy.
+type NumericIdentity uint32
+
+func ParseNumericIdentity(id string) (NumericIdentity, error) {
+	nid, err := strconv.ParseUint(id, 0, 32)
+	if err != nil {
+		return NumericIdentity(0), err
+	}
+	return NumericIdentity(nid), nil
+}
+
+func (id NumericIdentity) StringID() string {
+	return strconv.FormatUint(uint64(id), 10)
+}
+
+func (id NumericIdentity) String() string {
+	if v, exists := ReservedIdentityNames[id]; exists {
+		return v
+	}
+
+	return id.StringID()
+}
+
+// Uint32 normalizes the ID for use in BPF program.
+func (id NumericIdentity) Uint32() uint32 {
+	return uint32(id)
+}
+
+func GetReservedID(name string) NumericIdentity {
+	if v, ok := ReservedIdentities[name]; ok {
+		return v
+	}
+	return IdentityUnknown
+}
+
+// IsReservedIdentity returns whether id is one of the special reserved identities.
+func (id NumericIdentity) IsReservedIdentity() bool {
+	_, isReservedIdentity := ReservedIdentityNames[id]
+	return isReservedIdentity
+}
